tests/db_state_bench/db_state_zabacus_django/requirements.txt
aniso8601==3.0.2
django-cors-headers==2.4.0
django-extensions==2.2.5
django-graphql-jwt==0.1.13
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
mysqlclient==1.3.13
promise==2.1
PyJWT==1.6.4
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.3.0
typing==3.6.6
